Organize config utilities
