+++ src/applications/dependents/686c-674/config/utilities/index.js
@@ -1,0 +1,45 @@
+import {
+  isServerError,
+  isClientError,
+  getData,
+  checkAddingDependentsNot674ForPension,
+  checkAdding674ForPension,
+  showPensionBackupPath,
+  isVetInReceiptOfPension,
+  showPensionRelatedQuestions,
+  show674IncomeQuestions,
+  shouldShowStudentIncomeQuestions,
+} from './api';
+import {
+  validateName,
+  customFormReplacer,
+  buildSubmissionData,
+  customTransformForSubmit,
+  parseDateToDateObj,
+  spouseEvidence,
+  childEvidence,
+  showDupeModalIfEnabled,
+} from './data';
+
+export {
+  // api
+  isServerError,
+  isClientError,
+  getData,
+  checkAddingDependentsNot674ForPension,
+  checkAdding674ForPension,
+  showPensionBackupPath,
+  isVetInReceiptOfPension,
+  showPensionRelatedQuestions,
+  show674IncomeQuestions,
+  shouldShowStudentIncomeQuestions,
+  // data
+  validateName,
+  customFormReplacer,
+  buildSubmissionData,
+  customTransformForSubmit,
+  parseDateToDateObj,
+  spouseEvidence,
+  childEvidence,
+  showDupeModalIfEnabled,
+};
